Use an application truststore instead of the JDK
For this test, emptyStore.keystore is a jks keystore that contains only
a unused private key/public key pair.  There are no CA certificates of
any kind so it should fail any validation.

The WsWebSocketContainer uses the trust store and trust store password
properties from config user properties, not from system properties.  As
such I am creating them from the spring boot standard web socket client.
Since the trust store property must be a file object and since this code
is running in a jar, I cannot use an IDE/maven relative path for the
keystore.  The keystore must be packaged in the jar and at runtime
copied to a folder. The trust store property will then be set to that
temporary file location.

The temporary folder is created in the main method before starting the
spring boot and passed as an argument.